// sample/01-all-features/src/health/health.interface.ts
export interface DatabaseConnectionDetails {
    name: string;
    type: 'postgres';
    host: string;
    port: number;
    username: string;
    password: string;
    database: string;
  }
  
  export interface FusionAuthConnectionDetails {
    healthEndpoint: string;
  }
  
  export interface RedisConnectionDetails {
    host: string;
    port: number;
  }
  